refactor(email): depend on lexical_client directly
The port indirection existed only to dodge a dependency cycle:
`lexical_client -> models_search -> models_soup -> email`. That first edge was
a single 3-field DTO, `MarkdownParseResult`, describing a lexical service
response - so it belongs in the client, not in the search models.

Move it to `lexical_client::types`, drop `models_search` from the client, and
point its one other consumer (`search_processing_service`) at the new home.
`email` can then hold the client directly, and the `MarkdownEmailRenderer`
port and its `ai_tools` adapter go away.
