+++ crates/ai_tools/src/lib.rs
@@ -9,7 +9,6 @@ mod test;
 
 mod build_context;
 mod display_results;
-pub mod markdown_email_renderer;
 mod schemas;
 pub mod search;
 mod search_tools;
