+++ crates/email/Cargo.toml
@@ -46,6 +46,7 @@ attachment = { path = "../attachment", optional = true }
 anyhow = { workspace = true }
 async-trait = { workspace = true, optional = true }
 base64 = { workspace = true }
+lexical_client = { path = "../lexical_client" }
 entity_access = { path = "../entity_access" }
 entity_access_management = { path = "../entity_access_management", default-features = false, features = [
   "ports",
