+++ crates/lexical_client/Cargo.toml
@@ -7,7 +7,6 @@ version = "0.1.0"
 [dependencies]
 agent_fold = { path = "../agent_fold", default-features = false }
 anyhow = { workspace = true }
-models_search = { path = "../models_search" }
 reqwest = { workspace = true }
 serde = { workspace = true }
 tracing = { workspace = true }
